+++ example/lib/main.dart
@@ -16,7 +16,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
+  bool _timeSet = false;
 
   @override
   void initState() {
@@ -26,14 +26,13 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
-    String platformVersion;
+    bool isTimeAutomatic;
     // Platform messages may fail, so we use a try/catch PlatformException.
     // We also handle the message potentially returning null.
     try {
-      platformVersion =
-          await SystemTimeCheck.platformVersion ?? 'Unknown platform version';
+      isTimeAutomatic = await SystemTimeCheck.isSystemTimeAutomatic ?? false;
     } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
+      isTimeAutomatic = false;
     }
 
     // If the widget was removed from the tree while the asynchronous platform
@@ -42,7 +41,7 @@ class _MyAppState extends State<MyApp> {
     if (!mounted) return;
 
     setState(() {
-      _platformVersion = platformVersion;
+      _timeSet = isTimeAutomatic;
     });
   }
 
@@ -51,10 +50,10 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('is System time set?'),
         ),
         body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+          child: Text('Time is set: $_timeSet\n'),
         ),
       ),
     );