+++ lib/system_time_check.dart
@@ -1,4 +1,3 @@
-
 import 'dart:async';
 
 import 'package:flutter/services.dart';
@@ -6,8 +5,15 @@ import 'package:flutter/services.dart';
 class SystemTimeCheck {
   static const MethodChannel _channel = MethodChannel('system_time_check');
 
-  static Future<String?> get platformVersion async {
-    final String? version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
+  static Future<bool?> get isSystemTimeAutomatic async {
+    final bool? isSystemTimeAutomatic =
+        await _channel.invokeMethod('isSystemTimeAutomatic');
+    return isSystemTimeAutomatic;
+  }
+
+  static Future<bool?> get isSystemTimeAutomaticDebug async {
+    final bool? isSystemTimeAutomatic =
+        await _channel.invokeMethod('isSystemTimeAutomaticDebug');
+    return isSystemTimeAutomatic;
   }
 }